move signalr and management e2e test to standalone project (#579)
* move signalr and management e2e test to standalone project 

avoid HttpClientConnection version conflict

* remove unnecessary pkg
