+++ test/Microsoft.Azure.SignalR.E2ETests/SignalR/TestClientSet.cs
