+++ test/Microsoft.Azure.SignalR.E2ETests/SignalR/TestClientSetFactory.cs
