+++ test/Microsoft.Azure.SignalR.E2ETests/SignalR/TestStartup.cs
