+++ test/Microsoft.Azure.SignalR.Management.Tests/ServiceManagerFacts.cs
